[fix] 修复中值滤波窗口滑动问题

+++ filter.h
@@ -7,6 +7,6 @@ extern float filter_result_median;
 
 void filter_Median(float input, int filterwindowsize);
 
-bool compare_Median(int a, int b);
+bool compare_Median(float a, float b);
 
 #endif
